Refactor frontend to use vue-router so that we can have browser history and different views and components and other nice things
