+++ frontend/package.json
@@ -16,6 +16,7 @@
     "stylus": "^0.54.8",
     "stylus-loader": "^3.0.2",
     "vue": "^2.6.14",
+    "vue-router": "^3.5.2",
     "vuetify": "^2.5"
   },
   "devDependencies": {
