+++ frontend/src/components/Categories.vue
@@ -1,0 +1,8 @@
+<template>
+    <h1>THIS IS A TEST</h1>
+</template>
+<script>
+    export default {
+        
+    }
+</script>
